Do not log an error when PMD cannot analyze a file due to its content
PMD throws a PMDException when it cannot analyze a file. These
exceptions are logged as errors. Some of the causes for this exception
however are expected (e.g. a syntactically incorrect Java or XML file
that cannot be parsed) and are no longer logged as an error but as an
info log entry.