Improve path of a workspace relative rule set configuration
The path to a workspace relative rule set configuration is now
<project-name>/<project-relative-path> instead of just
<workspace-relative-path>. This way the path is independent of the
actual file system location of the workspace and a projects source code.
This allows one developer to checkout the projects directly into the
workspace while another to checkout the projects into a separate
directory outside of the workspace directory.