Move all Java related PMD support to a new plug-in.
This change moves all Java related support (e.g. quickfixes for Java
violations) to the newly created plug-in ch.acanda.eclipse.pmd.java.
This results in a core plug-in that does not depend on the JDT plug-ins.
So someone who uses PMD only for languages other than Java (e.g. XML or
JavaScript) does not have to install the JDT plug-ins.