Fix the "Unknown Language 'java'" message when building Java projects
The PMD languages are made available as services using the
java.util.ServiceLoader facility. The problem was that under some
circumstances the PMD languages were loaded with a class loader that
could not access the service implementations provided by the different
PMD jars. This fix ensures the services are loaded using a class loader
with access to the different service implementations (i.e. languages).

This fixes #21.